refactoring/user : service 인터페이스화 및 구현(Impl)
